Refactor: 메인, 랭킹 페이지에 쓰이는 타입을 따로 분리
