+++ src/features/main/RankingSection.tsx
@@ -1,56 +1,13 @@
 import TItem from "@/features/common/TItem";
 import { convertToURL } from "@/libs/utils/categoryMapper";
-import { ArtistRanking, TrackRanking } from "@prisma/client";
 import { useTranslation } from "next-i18next";
 import Link from "next/link";
 
-type RankingCategory =
-  | "ALL_TIME_ARTIST"
-  | "ALL_TIME_TRACK"
-  | "CURRENT_ARTIST"
-  | "CURRENT_TRACK";
-
-interface ArtistWithRanking
-  extends Omit<ArtistRanking, "updatedAt" | "rankingType"> {
-  rankingType: RankingCategory;
-  count: number;
-  updatedAt: string;
-  artist: {
-    id: number;
-    artistId: string;
-    name: string;
-    imageUrl: string;
-    followers: number;
-  };
-}
-
-interface TrackWithRanking
-  extends Omit<TrackRanking, "updatedAt" | "rankingType"> {
-  rankingType: RankingCategory;
-  count: number;
-  updatedAt: string;
-  track: {
-    id: number;
-    trackId: string;
-    name: string;
-    albumImageUrl: string;
-    artists: string;
-    popularity: number;
-  };
-}
-
-interface RankingSectionProps {
-  title: string;
-  data: ArtistWithRanking[] | TrackWithRanking[];
-  type: "artist" | "track";
-  category: RankingCategory;
-}
-
-const isArtistWithRanking = (
-  item: ArtistWithRanking | TrackWithRanking,
-): item is ArtistWithRanking => {
-  return (item as ArtistWithRanking).artist !== undefined;
-};
+import {
+  RankingSectionProps,
+  isArtistWithRanking,
+  isTrackWithRanking,
+} from "@/types/ranking";
 
 const RankingSection = ({
   title,
@@ -83,7 +40,7 @@ const RankingSection = ({
                   />
                 );
               }
-              if (type === "track" && !isArtistWithRanking(item)) {
+              if (type === "track" && isTrackWithRanking(item)) {
                 return (
                   <TItem
                     key={item.track.trackId}
