+++ src/pages/_app.tsx
@@ -1,5 +1,6 @@
 import Layout from "@/components/Layout";
 import ErrorBoundaryWrapper from "@/features/common/ErrorBoundaryWrapper";
+import GlobalLoadingBar from "@/features/common/GlobalLoadingBar";
 import CustomQueryClientProvider from "@/libs/react-query/CustomQueryClientProvider";
 import "@/styles/globals.css";
 import { SessionProvider } from "next-auth/react";
@@ -12,6 +13,7 @@ function App({ Component, pageProps: { session, ...pageProps } }: AppProps) {
       <CustomQueryClientProvider>
         <SessionProvider session={session}>
           <Layout>
+            <GlobalLoadingBar />
             <Component {...pageProps} />
           </Layout>
         </SessionProvider>
